move logic for preloading images to useFeed hook
