+++ packages/app/components/feed.tsx
@@ -132,7 +132,6 @@ const FollowingFeed = () => {
       {...queryState}
       bottomPadding={bottomBarHeight}
       data={queryState.data}
-      shouldPreloadImages={true}
     />
   );
 };
@@ -146,7 +145,6 @@ const AlgorithmicFeed = () => {
       {...queryState}
       bottomPadding={bottomBarHeight}
       data={queryState.data}
-      shouldPreloadImages={true}
     />
   );
 };
@@ -160,7 +158,6 @@ const CuratedFeed = () => {
       {...queryState}
       bottomPadding={safeAreaBottom}
       data={queryState.data}
-      shouldPreloadImages={true}
     />
   );
 };
